+++ src/main/java/com/divudi/bean/pharmacy/PharmacyItemExcelManager.java
@@ -674,7 +674,7 @@ public class PharmacyItemExcelManager implements Serializable {
                         amp = new Amp();
                         amp.setName(strAmp);
                         amp.setMeasurementUnit(strengthUnit);
-                        amp.setDblValue((float) strengthUnitsPerIssueUnit);
+                        amp.setDblValue((double) strengthUnitsPerIssueUnit);
                         amp.setCategory(cat);
                         amp.setVmp(vmp);
                         getAmpFacade().create(amp);
